[Unity][ONNX] Apply `R.nn.attention` in Onnx Attention OP (#62)
This PR refactors the frontend impl of Onnx attention op, to apply the
`R.nn.attention`. So we are able to use the cutlass BYOC with Onnx
frontend, or other backends for `R.nn.attention` in the future. And this
PR adds more support over Onnx attention op inputs and attributes.